Refactor entitlement tests to run faster.
