+++ packages/sdk/src/tests/multi/entitlements/spaceWithComplexEntitlements.test.ts
@@ -1,0 +1,346 @@
+/**
+ * @group with-entitlements
+ */
+
+import {
+    createTownWithRequirements,
+    createUserStreamAndSyncClient,
+    everyoneMembershipStruct,
+    expectUserCannotJoinSpace,
+    expectUserCanJoin,
+    linkWallets,
+    setupWalletsAndContexts,
+    twoNftRuleData,
+} from '../../testUtils'
+import { dlog } from '@river-build/dlog'
+import {
+    Address,
+    CheckOperationType,
+    LogicalOperationType,
+    Operation,
+    OperationType,
+    TestERC721,
+    treeToRuleData,
+    encodeThresholdParams,
+    createExternalNFTStruct,
+} from '@river-build/web3'
+
+const log = dlog('csb:test:spaceWithComplexEntitlements')
+
+describe('spaceWithComplexEntitlements', () => {
+    let testNft1Address: string, testNft2Address: string, testNft3Address: string
+    beforeAll(async () => {
+        ;[testNft1Address, testNft2Address, testNft3Address] = await Promise.all([
+            TestERC721.getContractAddress('TestNFT1'),
+            TestERC721.getContractAddress('TestNFT2'),
+            TestERC721.getContractAddress('TestNFT3'),
+        ])
+    })
+    test('two nft gate join pass', async () => {
+        const { alice, bob, aliceSpaceDapp, aliceProvider, alicesWallet, spaceId, channelId } =
+            await createTownWithRequirements({
+                everyone: false,
+                users: [],
+                ruleData: twoNftRuleData(testNft1Address, testNft2Address),
+            })
+
+        const aliceMintTx1 = TestERC721.publicMint('TestNFT1', alicesWallet.address as Address)
+        const aliceMintTx2 = TestERC721.publicMint('TestNFT2', alicesWallet.address as Address)
+
+        log('Minting nfts for alice')
+        await Promise.all([aliceMintTx1, aliceMintTx2])
+
+        log('Alice should be able to join space')
+        await expectUserCanJoin(
+            spaceId,
+            channelId,
+            'alice',
+            alice,
+            aliceSpaceDapp,
+            alicesWallet.address,
+            aliceProvider.wallet,
+        )
+
+        // kill the clients
+        const doneStart = Date.now()
+        await bob.stopSync()
+        await alice.stopSync()
+        log('Done', Date.now() - doneStart)
+    })
+
+    test('two nft gate join pass - across linked wallets', async () => {
+        const {
+            alice,
+            bob,
+            aliceSpaceDapp,
+            aliceProvider,
+            carolProvider,
+            alicesWallet,
+            carolsWallet,
+            spaceId,
+            channelId,
+        } = await createTownWithRequirements({
+            everyone: false,
+            users: [],
+            ruleData: twoNftRuleData(testNft1Address, testNft2Address),
+        })
+
+        const aliceMintTx1 = TestERC721.publicMint('TestNFT1', alicesWallet.address as Address)
+        const carolMintTx2 = TestERC721.publicMint('TestNFT2', carolsWallet.address as Address)
+
+        log('Minting nfts for alice and carol')
+        await Promise.all([aliceMintTx1, carolMintTx2])
+
+        log("linking carols wallet to alice's wallet")
+        await linkWallets(aliceSpaceDapp, aliceProvider.wallet, carolProvider.wallet)
+
+        log('Alice should be able to join space with one asset in carol wallet')
+        await expectUserCanJoin(
+            spaceId,
+            channelId,
+            'alice',
+            alice,
+            aliceSpaceDapp,
+            alicesWallet.address,
+            aliceProvider.wallet,
+        )
+
+        // kill the clients
+        const doneStart = Date.now()
+        await bob.stopSync()
+        await alice.stopSync()
+        log('Done', Date.now() - doneStart)
+    })
+
+    test('two nft gate join fail', async () => {
+        const { alice, bob, aliceSpaceDapp, aliceProvider, alicesWallet, spaceId } =
+            await createTownWithRequirements({
+                everyone: false,
+                users: [],
+                ruleData: twoNftRuleData(testNft1Address, testNft2Address),
+            })
+
+        // join alice
+        log('Minting an NFT for alice')
+        await TestERC721.publicMint('TestNFT1', alicesWallet.address as Address)
+
+        // first join the space on chain
+        const aliceJoinStart = Date.now()
+        log('transaction start Alice joining space')
+        const { issued } = await aliceSpaceDapp.joinSpace(
+            spaceId,
+            alicesWallet.address,
+            aliceProvider.wallet,
+        )
+        expect(issued).toBe(false)
+        log('Alice failed to join space and has a MembershipNFT', Date.now() - aliceJoinStart)
+
+        // Have alice create her own space so she can initialize her user stream.
+        // Then she will attempt to join the space from the client, which should fail.
+        await createUserStreamAndSyncClient(
+            alice,
+            aliceSpaceDapp,
+            'alice',
+            await everyoneMembershipStruct(aliceSpaceDapp, alice),
+            aliceProvider.wallet,
+        )
+        // Alice cannot join the space on the stream node.
+        await expectUserCannotJoinSpace(spaceId, alice, aliceSpaceDapp, alicesWallet.address)
+
+        // kill the clients
+        await bob.stopSync()
+        await alice.stopSync()
+    })
+
+    test('or of two nft gate join pass', async () => {
+        const { alice, bob, aliceSpaceDapp, aliceProvider, alicesWallet, spaceId, channelId } =
+            await createTownWithRequirements({
+                everyone: false,
+                users: [],
+                ruleData: twoNftRuleData(testNft1Address, testNft2Address, LogicalOperationType.OR),
+            })
+
+        // join alice
+        log('Minting an NFT for alice')
+        await TestERC721.publicMint('TestNFT1', alicesWallet.address as Address)
+
+        // first join the space on chain
+        log('Expect alice can join space')
+        await expectUserCanJoin(
+            spaceId,
+            channelId,
+            'alice',
+            alice,
+            aliceSpaceDapp,
+            alicesWallet.address,
+            aliceProvider.wallet,
+        )
+
+        // kill the clients
+        const doneStart = Date.now()
+        await bob.stopSync()
+        await alice.stopSync()
+        log('Done', Date.now() - doneStart)
+    })
+
+    test('or of two nft or one nft gate join pass', async () => {
+        const params = encodeThresholdParams({ threshold: 1n })
+        const leftOperation: Operation = {
+            opType: OperationType.CHECK,
+            checkType: CheckOperationType.ERC721,
+            chainId: 31337n,
+            contractAddress: testNft1Address as Address,
+            params,
+        }
+
+        const rightOperation: Operation = {
+            opType: OperationType.CHECK,
+            checkType: CheckOperationType.ERC721,
+            chainId: 31337n,
+            contractAddress: testNft2Address as Address,
+            params,
+        }
+        const two: Operation = {
+            opType: OperationType.LOGICAL,
+            logicalType: LogicalOperationType.AND,
+            leftOperation,
+            rightOperation,
+        }
+
+        const root: Operation = {
+            opType: OperationType.LOGICAL,
+            logicalType: LogicalOperationType.OR,
+            leftOperation: two,
+            rightOperation: {
+                opType: OperationType.CHECK,
+                checkType: CheckOperationType.ERC721,
+                chainId: 31337n,
+                contractAddress: testNft3Address as Address,
+                params,
+            },
+        }
+
+        const ruleData = treeToRuleData(root)
+        const { alice, bob, aliceSpaceDapp, aliceProvider, alicesWallet, spaceId, channelId } =
+            await createTownWithRequirements({
+                everyone: false,
+                users: [],
+                ruleData,
+            })
+
+        log("Mint Alice's NFTs")
+        const aliceMintTx1 = TestERC721.publicMint('TestNFT1', alicesWallet.address as Address)
+        const aliceMintTx2 = TestERC721.publicMint('TestNFT2', alicesWallet.address as Address)
+        await Promise.all([aliceMintTx1, aliceMintTx2])
+
+        log('expect alice can join space')
+        await expectUserCanJoin(
+            spaceId,
+            channelId,
+            'alice',
+            alice,
+            aliceSpaceDapp,
+            alicesWallet.address,
+            aliceProvider.wallet,
+        )
+
+        // kill the clients
+        const doneStart = Date.now()
+        await bob.stopSync()
+        await alice.stopSync()
+        log('Done', Date.now() - doneStart)
+    })
+
+    test('user with only one entitlement from 3-nested NFT rule data can join space', async () => {
+        const testNft1 = 'TestNft1'
+        const testNft2 = 'TestNft2'
+        const testNft3 = 'TestNft3'
+        const testNftAddress = await TestERC721.getContractAddress(testNft1)
+        const testNftAddress2 = await TestERC721.getContractAddress(testNft2)
+        const testNftAddress3 = await TestERC721.getContractAddress(testNft3)
+
+        const ruleData = createExternalNFTStruct([testNftAddress, testNftAddress2, testNftAddress3])
+
+        const {
+            alice,
+            bob,
+            carol,
+            aliceSpaceDapp,
+            aliceProvider,
+            carolProvider,
+            carolSpaceDapp,
+            alicesWallet,
+            carolsWallet,
+            spaceId,
+            channelId,
+        } = await createTownWithRequirements({
+            everyone: false,
+            users: [],
+            ruleData,
+        })
+        // Set up additional users to test single ownership of all three nfts.
+        const {
+            alice: dave,
+            alicesWallet: davesWallet,
+            aliceSpaceDapp: daveSpaceDapp,
+            aliceProvider: daveProvider,
+            carol: emily,
+            carolProvider: emilyProvider,
+            carolSpaceDapp: emilySpaceDapp,
+            carolsWallet: emilyWallet,
+        } = await setupWalletsAndContexts()
+
+        // Have alice create her own space so she can initialize her user stream.
+        // Then she will attempt to join the space from herclient, which should fail.
+        await createUserStreamAndSyncClient(
+            alice,
+            aliceSpaceDapp,
+            'alice',
+            await everyoneMembershipStruct(aliceSpaceDapp, alice),
+            aliceProvider.wallet,
+        )
+        await expectUserCannotJoinSpace(spaceId, alice, aliceSpaceDapp, alicesWallet.address)
+
+        await TestERC721.publicMint(testNft1, carolsWallet.address as Address)
+        await expectUserCanJoin(
+            spaceId,
+            channelId,
+            'carol',
+            carol,
+            carolSpaceDapp,
+            carolsWallet.address,
+            carolProvider.wallet,
+        )
+
+        await TestERC721.publicMint(testNft2, davesWallet.address as Address)
+        await expectUserCanJoin(
+            spaceId,
+            channelId,
+            'dave',
+            dave,
+            daveSpaceDapp,
+            davesWallet.address,
+            daveProvider.wallet,
+        )
+
+        await TestERC721.publicMint(testNft3, emilyWallet.address as Address)
+        await expectUserCanJoin(
+            spaceId,
+            channelId,
+            'emily',
+            emily,
+            emilySpaceDapp,
+            emilyWallet.address,
+            emilyProvider.wallet,
+        )
+
+        // kill the clients
+        const doneStart = Date.now()
+        await bob.stopSync()
+        await alice.stopSync()
+        await carol.stopSync()
+        await dave.stopSync()
+        await emily.stopSync()
+        log('Done', Date.now() - doneStart)
+    })
+})
