+++ packages/sdk/src/tests/multi_v2/entitlements/channelsWithErc1155Entitlement.test.ts
@@ -9,15 +9,14 @@ import {
     expectUserCanJoinChannel,
     expectUserCannotJoinChannel,
     linkWallets,
-    mockCrossChainCheckOp,
-} from '../testUtils'
+} from '../../testUtils'
 import { dlog } from '@river-build/dlog'
-import { Address, treeToRuleData, TestERC1155, TestCrossChainEntitlement } from '@river-build/web3'
+import { Address, treeToRuleData, TestERC1155 } from '@river-build/web3'
 
-const log = dlog('csb:test:channelsWithV2Entitlements')
+const log = dlog('csb:test:channelsWithErc1155Entitlements')
 const test1155Name = 'TestERC1155'
 
-describe('channelsWithV2Enitlements', () => {
+describe('channelsWithErc1155Entitlements', () => {
     test('erc1155 gate join pass', async () => {
         const ruleData = treeToRuleData(
             await erc1155CheckOp(test1155Name, BigInt(TestERC1155.TestTokenId.Bronze), 1n),
@@ -63,7 +62,7 @@ describe('channelsWithV2Enitlements', () => {
         log('Done', Date.now() - doneStart)
     })
 
-    test('ERC1155 gate join pass - join as root, asset in linked wallet', async () => {
+    test('erc1155 gate join pass - join as root, asset in linked wallet', async () => {
         const ruleData = treeToRuleData(
             await erc1155CheckOp(test1155Name, BigInt(TestERC1155.TestTokenId.Bronze), 1n),
         )
@@ -105,7 +104,7 @@ describe('channelsWithV2Enitlements', () => {
         log('Done', Date.now() - doneStart)
     })
 
-    test('ERC1155 Gate Join Pass - join as linked wallet, assets in root wallet', async () => {
+    test('erc1155 Gate Join Pass - join as linked wallet, assets in root wallet', async () => {
         const ruleData = treeToRuleData(
             await erc1155CheckOp(test1155Name, BigInt(TestERC1155.TestTokenId.Bronze), 1n),
         )
@@ -142,7 +141,7 @@ describe('channelsWithV2Enitlements', () => {
         log('Done', Date.now() - doneStart)
     })
 
-    test('ERC1155 Gate Join Pass - assets split across wallets', async () => {
+    test('erc1155 Gate Join Pass - assets split across wallets', async () => {
         const ruleData = treeToRuleData(
             await erc1155CheckOp(test1155Name, BigInt(TestERC1155.TestTokenId.Bronze), 2n),
         )
@@ -174,138 +173,6 @@ describe('channelsWithV2Enitlements', () => {
         )
 
         log('expect that alice can join the space')
-        // Validate alice can join the channel
-        await expectUserCanJoinChannel(alice, aliceSpaceDapp, spaceId, channelId!)
-
-        const doneStart = Date.now()
-        // kill the clients
-        await bob.stopSync()
-        await alice.stopSync()
-        log('Done', Date.now() - doneStart)
-    })
-
-    test('cross chain entitlement gate pass', async () => {
-        const idParam = 1n
-        const contractName = 'TestCrossChain'
-        const ruleData = treeToRuleData(await mockCrossChainCheckOp(contractName, idParam))
-
-        const { alice, bob, alicesWallet, aliceSpaceDapp, spaceId, channelId } =
-            await setupChannelWithCustomRole([], ruleData)
-
-        await TestCrossChainEntitlement.setIsEntitled(
-            contractName,
-            alicesWallet.address as Address,
-            idParam,
-            true,
-        )
-
-        log('expect that alice can join the channel')
-        await expectUserCanJoinChannel(alice, aliceSpaceDapp, spaceId, channelId!)
-
-        // kill the clients
-        const doneStart = Date.now()
-        await bob.stopSync()
-        await alice.stopSync()
-        log('Done', Date.now() - doneStart)
-    })
-
-    test('cross chain entitlement gate fail', async () => {
-        const idParam = 1n
-        const contractName = 'TestCrossChain'
-        const ruleData = treeToRuleData(await mockCrossChainCheckOp(contractName, idParam))
-
-        const { alice, bob, alicesWallet, aliceSpaceDapp, spaceId, channelId } =
-            await setupChannelWithCustomRole([], ruleData)
-
-        await TestCrossChainEntitlement.setIsEntitled(
-            contractName,
-            alicesWallet.address as Address,
-            idParam,
-            false,
-        )
-
-        log('expect that alice cannot join the channel')
-        await expectUserCannotJoinChannel(alice, aliceSpaceDapp, spaceId, channelId!)
-
-        // kill the clients
-        const doneStart = Date.now()
-        await bob.stopSync()
-        await alice.stopSync()
-        log('Done', Date.now() - doneStart)
-    })
-
-    test('cross chain entitlement gate join pass - join as root, linked wallet entitled', async () => {
-        const idParam = 1n
-        const contractName = 'TestCrossChain'
-        const ruleData = treeToRuleData(await mockCrossChainCheckOp(contractName, idParam))
-
-        const {
-            alice,
-            bob,
-            aliceSpaceDapp,
-            aliceProvider,
-            carolsWallet,
-            carolProvider,
-            spaceId,
-            channelId,
-        } = await setupChannelWithCustomRole([], ruleData)
-
-        // Link carol's wallet to alice's as root
-        await linkWallets(aliceSpaceDapp, aliceProvider.wallet, carolProvider.wallet)
-
-        log('expect that alice cannot join the channel')
-        await expectUserCannotJoinChannel(alice, aliceSpaceDapp, spaceId, channelId!)
-
-        // Set entitlement for carol's wallet
-        await TestCrossChainEntitlement.setIsEntitled(
-            contractName,
-            carolsWallet.address as Address,
-            idParam,
-            true,
-        )
-
-        // Wait 2 seconds for the negative auth cache to expire
-        await new Promise((f) => setTimeout(f, 2000))
-
-        // Validate alice can join the channel
-        await expectUserCanJoinChannel(alice, aliceSpaceDapp, spaceId, channelId!)
-
-        const doneStart = Date.now()
-        // kill the clients
-        await bob.stopSync()
-        await alice.stopSync()
-        log('Done', Date.now() - doneStart)
-    })
-
-    test('cross chain entitlement gated join - join as linked wallet, assets in root wallet', async () => {
-        const idParam = 1n
-        const contractName = 'TestCrossChain'
-        const ruleData = treeToRuleData(await mockCrossChainCheckOp(contractName, idParam))
-
-        const {
-            alice,
-            bob,
-            aliceSpaceDapp,
-            carolSpaceDapp,
-            aliceProvider,
-            carolsWallet,
-            carolProvider,
-            spaceId,
-            channelId,
-        } = await setupChannelWithCustomRole([], ruleData)
-
-        log("Joining alice's wallet as a linked wallet to carol's root wallet")
-        await linkWallets(carolSpaceDapp, carolProvider.wallet, aliceProvider.wallet)
-
-        // Set carol's wallet as entitled
-        await TestCrossChainEntitlement.setIsEntitled(
-            contractName,
-            carolsWallet.address as Address,
-            idParam,
-            true,
-        )
-
-        log('expect that alice can join the channel')
         // Validate alice can join the channel
         await expectUserCanJoinChannel(alice, aliceSpaceDapp, spaceId, channelId!)
